refactor: extract class `RootPreferencesFragment` from `SettingsActivity` and move class into `fragment.settings`
` SettingsActivity` class does the work of two. so the other class need to be extracted. and the place is  more suitable in `fragments.settings`  than  `activities`
